moved inmem to mod
