+++ server/src/main.rs
@@ -1,33 +1,9 @@
-use sam_server::{
-    managers::in_memory::{
-        account::InMemoryAccountManager, device::InMemoryDeviceManager, keys::InMemoryKeyManager,
-        message::InMemoryMessageManager,
-    },
-    start_server,
-    state::{state_type::StateType, ServerState},
-    ServerConfig,
-};
-
-struct InMemStateType;
-
-impl StateType for InMemStateType {
-    type AccountManager = InMemoryAccountManager;
-    type DeviceManager = InMemoryDeviceManager;
-    type MessageManager = InMemoryMessageManager;
-    type KeyManager = InMemoryKeyManager;
-}
+use sam_server::{start_server, state::ServerState, ServerConfig};
 
 #[tokio::main]
 pub async fn main() {
-    let state: ServerState<InMemStateType> = ServerState::new(
-        InMemoryAccountManager::new(),
-        InMemoryDeviceManager::new("secret".to_string()),
-        InMemoryMessageManager::new(),
-        InMemoryKeyManager::new(),
-    );
-
     let config = ServerConfig {
-        state,
+        state: ServerState::in_memory_default("test".to_string()),
         addr: "127.0.0.1:8080"
             .parse()
             .expect("Unable to parse socket address"),
